Rename from state to reducer
